requirements.txt
django==5.1.7
djangorestframework==3.15.2
psycopg2-binary==2.9.10
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.9.0
django-cors-headers==4.7.0
django-allauth==65.7.0
Pillow==11.2.1
django-filter==25.1.0